admin-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.35.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
google-api-python-client==2.147.0
google-auth==2.35.0
pydantic-settings==2.5.2
pydantic==2.9.2
